requirements.txt
streamlit>=1.31.0
crewai>=0.28.0
crewai-tools>=0.1.6
langchain-community>=0.0.29
langchain-groq>=0.1.3
langchain-openai>=0.1.7
sentence-transformers>=2.3.1
tavily-python>=0.3.0
python-dotenv>=1.0.1
qdrant-client>=1.15.0
pypdf>=3.17.4
